Add a 'meson' version scheme comparison
Use it for meson version comparisons (in project(meson_version:), and for
FeatureNew checks)

Note that when comparing against coredata.version, we need to be able to
relax strict validation of the version string (may consist of only digits
and '.') for development versions (e.g. '0.48.0.dev1') to be considered a
valid version string.