Simplify get_methods()
At the moment, a Dependency subclass's get_methods() method is used by
_process_method_kw() to define both (a) the default list of methods to try
for method:auto, and (b) the valid set of methods.

Having the set of valid methods vary depending on the host machine is a bit
odd, as it makes things like:

dependency('python3', method: 'sysconfig')
dependency('cups', method: 'extraframework')
dependency('gl', method: 'system')
dependency('gnustep', method: 'extraframework')

valid or invalid depending on the host machine.

(Further, at the moment it actually depends on the build machine, which
seems incorrect, and that is somewhat tricky to fix)

Change so it only defines (b) the valid set of methods, and ensure that
those previously invalid methods always fail for host machines they cannot
work on, so they are safe to include in the default list.