Factor out invoking compiler to link to a separate class
Currently the linker syntax is assumed to be fixed per-compiler.  Factor
this out so we can support compilers which may be used with multiple
possible linkers with different syntax.

Usually, we invoke a compiler driver for a final link, as it has
knowledge of standard libraries to use etc., but in some cases we use
the linker.

Also drop redundant checking of type of vs_module_defs:, it been done in
the interpreter, as it should be, for a while now.