Factor out invoking compiler to link to a separate class
Usually, we invoke a compiler driver for a final link, as it has
knowledge of standard libraries to use etc., but in some cases we use
the linker.

(MSVC embeds 'linker directives' into the output objects which tell the
linker what standard libraries are needed)

Also drop redundant checking of type of vs_module_defs:, it been done in
the interpreter, as it should be, for a while now.

(This is needed because the linker that clang invokes varies depending
on the target, and so requires different syntax)