ninja: Quoting in rspfile depends on the compiler, not the shell
In certain exotic configurations, the style of quoting expected in the
response file may not match that expected by the shell

e.g. under MSYS2, ninja invokes commands via CreateProcess (which
results in cmd-style quoting processed by parse_cmdline or
CommandLineToArgvW), but gcc will use sh-style quoting in any response
file it reads.

Since we now properly sh-quote gcc responsefiles, there is no need for
the 'compiler-specific escaping' in escape_extra_args() to do that (for
-D arguments only, but since we don't use Windows-style paths anywhere,
that's the only place we expected to see them)