Redetect machines after every change to languages
Currently, detect_machine_info() is called:

- from Environment.__init__(), before we have any compiler available

- from Interpreter.__init__() with the list of languages provided to
project() (determined via the initial parse_project())

There is a problem with compiler-informed machine detection only
happening once, immediately after project():

If we are building on a x86_64 build machine, but the host compiler is
x86, we can't notice that the host machine is x86 until after we are
told to use that compiler.

Example: MSYS2 64-bit, using the "MingW-w64 Win32 Shell", which is
64-bit, but has a 'i686-w64-mingw32' targeted gcc as 'gcc'.

'test/common/123 cpp and asm' fails, as it (incorrectly) tries to build
retval-x86_64.S using an x86 compiler.

Redetect machines after project(), and after every add_languages().