Ensure MESON_CI_JOBNAME is set for all CI unittest runs
Check that MESON_CI_JOBNAME is set in all CI runs which run unittests.
Set MESON_CI_JOBNAME for GitHub workflows which run unittests.

(Note that ${{ github.job }} is the literal job.id used in the yaml, not
any name given to the job, and hence is the same for all matrix entries)